Dashboard + all modules: modules now use module.q
module is passed in the init function and has the .q and .qAll functions that replace document.querySelector{All} functions to allow modules to handle better when multiple instances of them exist in the layout.

module ~should~ also be used to store local variables for a specific instance of the module. However, the only module that this would apply to currently is multiTimer, and it is not supported currently.