Multitimer: Fixed createWith
createWith was a way to create a timer, meant to be used with the eval function of codeEditor. It now functions with the new changes to modules instances.
This function has also been added to the documentation.
This comes with the change that modules now store an array of instances of that module on the template.