Dashboard: Added option to default settings
This option allows the user to have their settings auto change if a default setting is changed at a later date. This also saves some space in localStorage as the number of settings grows. A plus side of this is that it also removes the need for the message that prompts you to save the settings for the first time.