Dashboard: Add config to the settings
The config logic existed prior to the settings pane, this moves it there. The text type isn't the best for this, ideally this would have some sort of custom UI, but this at least makes it possible to change this. If the layout is messed up, the settings pane still works to revert changes.