multitimer: Add a sort button for the timers
The sort button will sort timers with priority ringing; lowest number first, unstarted, active/paused; lowest number first. This is an approximation to the priority that someone might want to deal with the timers.

While doing this, I ran into a quirk with multitimer where it doesn't work if you save before loading the timers because ringing timers rely on being unsaved. This will cause an issue if someone changes a timer while one is ringing and it saves as reset, but as that is an unlikely occurrence and the entire multitimer module is due for a rewrite, I'm going to fix the issue in an improper way for now.