multitimer: Save the status in numeric format
Previously it was using the value of the start button, after the rewrite this is not what is being used internally. So it makes sense to save it as a the number that is now used elsewhere. This also renamed state to status to remain consistent across the save data and the internal variables.