all: Module specific CSS
Each module will now has a CSS class of the modules name used for applying rules to only the elements within that module. This should make it less likely for modules to unknowingly rely on each other or conflict with each other.

This also removed the half-followed convention of putting a two letter abbreviation before classes.