todo: Don't type a newline when saving description
Previously if you hit enter in the description it would type an enter character and then bubble the event to the save listener.
Now if you hit enter and are not holding shift, it will not type the newline and will only bubble to the  event listener without editing the description.