multitimer: Enter overriding focused element
Hitting enter anywhere on the element would previously trigger the start button, this was not ideal because if the start button was focused it would double hit it. This change makes it so that you must be focused on either the name or the time entry to trigger it. This also fixes a bug where you could not tab to the reset button.