dashboard: Add tests for every module
Add a new concept in the tests folder, all modules
have a test/*_tests.js file that is loaded and
called by dashboard when tests are enabled. Tests
are enabled by changing the code to set
dashboard.tests.enable=true to prevent the tests
being run on real data. This includes replacing
all calls to alert() and confirm() to new
functions in common.js to avoid making the user
have to to enter through these for the tests.
Tests are  run in a single layout with no other
modules to prevent interference. Tests have access
to the dashboard.tests.click() function to
synchronously run a click event to simulate user
interaction. They can also run
dashboard.tests.forceSetting() to override a
setting for the remainder of the session.