All: Standardize use of module and instance
Module is now the template for creating the instances, instances are the object that defines an occurrence of the module as well as the DOM elements that are associated with it.

This should prevent code where both modules and instances are used in the same area and the variable names made no sense.