dashboard: Add a GUI config editor
This gives the user the option to edit their config without needing to know the JSON structure. There is still an option to edit the config as a string, mainly for copying and pasting if needed.

For this change to happen, settings now have a way to provide a custom JS function to run to display on the settings page. Once this function is completed, it should move a string to a hidden input element so that the settings manager can actually use it.

Also change the maxHeight to a "name" "value" pair instead of just being it's own tag, since that ends up being easier to parse in most cases.

The container and instance settings have been reworked to be more modular to reduce the number of places that have to be edited to add or change them.