dashboard: Do updates for when layout is overriden
Previously, update functions were being called as part of the page layout, and also on page load. I believe the intention there was to only do the updates for modules that were in the current layout. The flaw with this was that overrideLayout modules are not on the layout. This mostly worked because the update function was specifically invoked for all modules on page load, but not when loading. To fix this, only call the update function on every module both when loading and opening the page.